+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## NEXT
+
+- Feat: Update Plugin SDK to latest version
+- Fix: Use configured limit for sync action
+- Fix: Missing files are deleted if activated in the settings
+- Add documentation for sync command
+
 ## 2.3.0
 
 - Fix file name collision. When a file was added with the same file name, the import was skipped.